changed nft value to float type

+++ packages/std/src/helpers.rs
@@ -21,7 +21,7 @@ pub struct TransferNft{
     pub comment:    String, 
     pub nft:        String,
     pub nft_data:   String,
-    pub nft_value:  i32,
+    pub nft_value:  f64,
     pub owner:      String,
     pub quorum_type:  i32,
     pub receiver:    String,
